Extract `printDecorators`
